Combine the separate test directories
