+++ .github/workflows/regression-tests-and-codegen.yaml
@@ -187,7 +187,7 @@ jobs:
         env:
           KUBE2E_TESTS: discovery
         run: |
-          make run-kind-ci-regression-tests
+          make run-ci-regression-tests
       - name: Debug Info
         if: failure()
         run: |
