+++ test/kube2e/discovery/README.md
