+++ test/kube2e/utils/curl.go
